Reorganize and hide unused stories
